Move `UntypedHandle` to separate module
